requirements.txt
Django==4.2.13
djangorestframework==3.15.1
Pillow==10.3.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn
django-filter

